extract db types to separate file
